fix(parental-leave): Remove SummaryTimeline
Removed SummaryTimeline and added it directly to components that were
using it

UUFV-394
