+++ libs/application/templates/parental-leave/src/fields/EditOrAddEmployersAndPeriodsReview/review-groups/Periods.tsx
@@ -1,7 +1,10 @@
 import { Application } from '@island.is/application/types'
-import { ReviewGroup } from '@island.is/application/ui-components'
-import React, { FC } from 'react'
-import { SummaryTimeline } from '../../components/SummaryTimeline/SummaryTimeline'
+import { DataValue, ReviewGroup } from '@island.is/application/ui-components'
+import { Box } from '@island.is/island-ui/core'
+import { useLocale } from '@island.is/localization'
+import { FC } from 'react'
+import { parentalLeaveFormMessages } from '../../../lib/messages'
+import { formatPeriods } from '../../../lib/parentalLeaveUtils'
 
 interface ReviewScreenProps {
   application: Application
@@ -12,13 +15,42 @@ const Periods: FC<React.PropsWithChildren<ReviewScreenProps>> = ({
   application,
   goToScreen,
 }) => {
+  const { formatMessage, formatDateFns } = useLocale()
+  const periods = formatPeriods(application, formatMessage)
   return (
     <ReviewGroup
       isEditable
       editAction={() => goToScreen?.('addPeriods')}
       isLast
     >
-      <SummaryTimeline application={application} />
+      <Box>
+        {periods.map((period, index) => {
+          const value = period.actualDob
+            ? formatMessage(
+                parentalLeaveFormMessages.reviewScreen.periodActualDob,
+                {
+                  duration: period.duration,
+                },
+              )
+            : `${formatDateFns(period.startDate)} — ${formatDateFns(
+                period.endDate,
+              )}`
+
+          return (
+            <DataValue
+              key={`SummaryTimeline-${index}`}
+              label={formatMessage(
+                parentalLeaveFormMessages.reviewScreen.period,
+                {
+                  index: index + 1,
+                  ratio: period.ratio,
+                },
+              )}
+              value={value}
+            />
+          )
+        })}
+      </Box>
     </ReviewGroup>
   )
 }
